requirements.txt
amqp==5.0.6
asgiref==3.3.4
astroid==2.4.2
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
cryptography==3.4.7
Django==3.2
django-celery-results==2.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
humanize==3.4.1
idna==2.10
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.2
Pillow==8.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.18
pycparser==2.20
pylint==2.6.0
redis==3.5.3
requests==2.25.1
toml==0.10.1
tornado==6.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.12.1